refactor(auth): divide into subunits
