+++ src/auth/auth.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
-import { AuthService } from './auth.service';
-import { VerifyLoginService } from './verify-login/verify-login.service';
+import { AuthService } from './services/auth.service';
+import { VerifyLoginService } from './services/verify-login.service';
 import { AuthController } from './auth.controller';
 import { ConfigModule } from '@nestjs/config';
-import { JwtAccessStrategy } from './jwt-access.strategy';
-import { JwtRefreshStrategy } from './jwt-refresh.strategy';
+import { JwtAccessStrategy } from './strategies/jwt-access.strategy';
+import { JwtRefreshStrategy } from './strategies/jwt-refresh.strategy';
 import { DatabaseModule } from '../database/database.module';
 
 @Module({
