+++ src/auth/tests/functional.services/auth.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AuthService } from './auth.service';
+import { AuthService } from '../../services/auth.service';
 
 describe('AuthService', () => {
     let service: AuthService;
