+++ src/auth/tests/integration.controller/auth.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { AuthController } from '../../auth.controller';
+
+describe('AuthController', () => {
+    let controller: AuthController;
+
+    beforeEach(async () => {
+        const module: TestingModule = await Test.createTestingModule({
+            controllers: [AuthController],
+        }).compile();
+
+        controller = module.get<AuthController>(AuthController);
+    });
+
+    it('should be defined', () => {
+        expect(controller).toBeDefined();
+    });
+});
